check norm, header, function name

+++ 42cursus/libft/ft_strtrim.c
@@ -6,12 +6,11 @@
 /*   By: si-park <si-park@student.42gyeongsan.kr    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 17:19:11 by si-park           #+#    #+#             */
-/*   Updated: 2024/11/10 01:04:48 by si-park          ###   ########.fr       */
+/*   Updated: 2024/11/11 16:13:30 by si-park          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-#include <stdlib.h>
 
 static int	is_in_set(char c, char *set)
 {
@@ -26,10 +25,10 @@ static int	is_in_set(char c, char *set)
 
 char	*ft_strtrim(char const *s1, char const *set)
 {
-	int		i;
-	int		start;
-	int		end;
-	char	*trimmed_str;
+	size_t		i;
+	size_t		start;
+	size_t		end;
+	char		*trimmed_str;
 
 	start = 0;
 	end = ft_strlen(s1) - 1;
